% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{lifecycle}
\alias{lifecycle}
\title{Life cycle of the shredder package}
\description{
\lifecycle{maturing}

The shredder package is currently maturing. Unless otherwise stated,
this applies to all its exported functions. Maturing functions are
susceptible to API changes. Only use these in packages if you're
prepared to make changes as the package evolves. See sections below
for a list of functions marked as stable.

The documentation pages of retired functions contain life cycle
sections that explain the reasons for their retirements.
}
\section{Stable functions}{


\lifecycle{stable}
\itemize{
\item \code{\link[=stan_retain]{stan_retain()}}
\item \code{\link[=stan_select]{stan_select()}}
\item \code{\link[=stan_contains]{stan_contains()}}, \code{\link[=stan_starts_with]{stan_starts_with()}}, \code{\link[=stan_ends_with]{stan_ends_with()}}
\item \code{\link[=stan_names]{stan_names()}}
\item \code{\link[=stan_slice]{stan_slice()}}
\item \code{\link[=stan_sample_n]{stan_sample_n()}}, \code{\link[=stan_sample_frac]{stan_sample_frac()}}
}

\lifecycle{maturing}
\itemize{
\item \code{\link[=stan_filter]{stan_filter()}}
}
}

\keyword{internal}
