% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{peek_pars}
\alias{peek_pars}
\title{Access stored pars for stan partials}
\usage{
peek_pars()
}
\value{
character
}
\description{
Access pars currently stored that \link[shredder:stan_partials]{stan_partials} use.
}
