% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats_example.R
\name{rats_example}
\alias{rats_example}
\title{Rats model}
\usage{
rats_example()
}
\value{
stanfit object
}
\description{
Run Rats data model
}
\examples{
set.seed(123)
rats <- rats_example()
rats
}
\seealso{
\link[rstan:stan]{stan}
\link[parallel:detectCores]{detectCores}
}
\concept{examples}
