% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_axe.R
\name{stan_axe}
\alias{stan_axe}
\title{Remove elements from stanfit object}
\usage{
stan_axe(object, what = c("summary", "fit_instance", "stanmodel"))
}
\arguments{
\item{object}{stanfit}

\item{what}{name of attribute to remove, Default: c("summary", "fit_instance", "stanmodel")}
}
\value{
stanfit
}
\description{
Safely remove large elements from stanfit objects that are not needed for postprocessing.
}
\details{
shredder can remove three elements in the \code{stanfit} object
\itemize{
\item The cached fit summary stored in fit@\code{.MISC}$summary
\item The cached cpp object stored in fit@.MISC$stan_fit_instance
\item The stanmodel stored in fit@stanmodel
}
}
\seealso{
Other utility: 
\code{\link{stan_names}()},
\code{\link{stan_retain}()}
}
\concept{utility}
