% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_filter.R
\name{stan_filter}
\alias{stan_filter}
\title{FUNCTION_TITLE}
\usage{
stan_filter(object, ..., chain = 1)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{chain}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{quo_squash}},\code{\link[rlang]{quotation}}
\code{\link[stringi]{stri_extract_all}}
\code{\link[purrr]{keep}},\code{\link[purrr]{flatten}},\code{\link[purrr]{map}}
}
