% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_keep.R
\name{stan_keep}
\alias{stan_keep}
\title{Return object with specific chains}
\usage{
stan_keep(object, chains = 1)
}
\arguments{
\item{object}{stanfit object}

\item{chains}{numeric, chains to keep, Default: 1}
}
\value{
stanfit object
}
\description{
Use stan_keep() to choose chainsto preserve.
}
\examples{
rats <- rats_example(nCores = 1)

#keep first chain only
rats\%>\%
  stan_keep()
  
#keep chains 1 and 3
rats\%>\%
  stan_keep(c(1,3))
  
}
\seealso{
Other utility: \code{\link{stan_names}}
}
\concept{utility}
