% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_names.R
\name{stan_names}
\alias{stan_names}
\title{The Par Names of a Stanfit Object}
\usage{
stan_names(x, expand = FALSE)
}
\arguments{
\item{x}{a stanfit object}

\item{expand}{logical, if TRUE par names are returned including dimesnion
indicies (fnames_oi), Default: FALSE}
}
\value{
character
}
\description{
Functions to get the names of a stanfit object.
}
\examples{
rats <- rats_example(nCores = 1)

rats\%>\%stan_names()

rats\%>\%stan_names(expand = TRUE)

}
\seealso{
Other utility: 
\code{\link{stan_axe}()},
\code{\link{stan_retain}()}
}
\concept{utility}
