% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{stan_contains}
\alias{stan_contains}
\alias{stan_starts_with}
\alias{stan_ends_with}
\title{Partial match functions for stanfit objects}
\usage{
stan_contains(match, ignore.case = TRUE, pars = peek_pars())

stan_starts_with(match, ignore.case = TRUE, pars = peek_pars())

stan_ends_with(match, ignore.case = TRUE, pars = peek_pars())
}
\arguments{
\item{match}{character, pattern to search for}

\item{ignore.case}{logical to ignore the case, Default: TRUE}

\item{pars}{list of pars to search in, Default: peek_pars()}
}
\value{
character
}
\description{
Partial match functions used in \link[=stan_select]{stan_select}
}
\examples{
\donttest{
rats <- rats_example()

rats\%>\%
  stan_names()

rats\%>\%
  stan_select(stan_contains('sq'))

rats\%>\%
  stan_select(mu_alpha, stan_contains('sq'))

 
rats\%>\%
  stan_select(stan_starts_with('mu'))

rats\%>\%
  stan_select(stan_starts_with('mu'),stan_contains('sq'))

rats\%>\%
  stan_select(stan_ends_with('0'))

# free form regex in contains   
rats\%>\%
  stan_select(stan_contains('sq|mu'))
}

}
\seealso{
Other subsetting: 
\code{\link{stan_select}()}
}
\concept{subsetting}
