% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sample_frac.R, R/stan_sample_n.R
\name{stan_sample_frac}
\alias{stan_sample_frac}
\alias{stan_sample_n}
\title{Sample n post-warumps samples from a stanfit object}
\usage{
stan_sample_frac(object, size)

stan_sample_n(object, size)
}
\arguments{
\item{object}{stanfit object}

\item{size}{numeric, for \link[=stan_sample_n]{shredder} size of sample, for
\link[=stan_frac]{shredder} fraction of samples to sample.}
}
\value{
stanfit
}
\description{
This is a wrapper around sample.int() to make it easy to select
random samples from each parameter in a stanfit object.
}
\examples{
set.seed(123)

rats <- rats_example()

rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_sample_n(30)
  
}
\seealso{
\link[purrr:map]{map}

Other filtering: \code{\link{stan_filter}},
  \code{\link{stan_slice}}
}
\concept{filtering}
