% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sample_frac.R
\name{stan_sample_frac}
\alias{stan_sample_frac}
\title{FUNCTION_TITLE}
\usage{
stan_sample_frac(object, size)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{size}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
