% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_select.R
\name{stan_select}
\alias{stan_select}
\title{FUNCTION_TITLE}
\usage{
stan_select(object, ...)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{quotation}},\code{\link[rlang]{quo_label}}
\code{\link[purrr]{map}}
}
