% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{stan_contains}
\alias{stan_contains}
\alias{stan_starts_with}
\alias{stan_ends_with}
\title{Partial match functions for stanfit objects}
\usage{
stan_contains(match, ignore.case = TRUE)

stan_starts_with(match, ignore.case = TRUE)

stan_ends_with(match, ignore.case = TRUE)
}
\arguments{
\item{match}{character, pattern to search for}

\item{ignore.case}{logical to ignore the case, Default: TRUE}
}
\value{
character
}
\description{
Partial match functions used in stan_select
}
