% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_slice.R
\name{stan_slice}
\alias{stan_slice}
\title{FUNCTION_TITLE}
\usage{
stan_slice(object, ..., inc_warmup = TRUE)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}

\item{inc_warmup}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
