% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_split.R
\name{stan_split}
\alias{stan_split}
\title{FUNCTION_TITLE}
\usage{
stan_split(object, ncut = 10, inc_warmup = TRUE)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{ncut}{PARAM_DESCRIPTION, Default: 10}

\item{inc_warmup}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
}
