% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_thin.R
\name{stan_thin_n}
\alias{stan_thin_n}
\alias{stan_thin_frac}
\title{thin n post-warumps samples from a stanfit object}
\usage{
stan_thin_n(object, size, inc_warmup = TRUE)

stan_thin_frac(object, size, inc_warmup = TRUE)
}
\arguments{
\item{object}{stanfit object}

\item{size}{numeric, for \link[shredder:stan_thin_n]{stan_thin_n} size of thin, for
\link[shredder:stan_thin_frac]{stan_thin_frac} fraction of samples to sample.}

\item{inc_warmup}{logical, include warmup in output, Default: TRUE}
}
\value{
stanfit
}
\description{
This is a wrapper around stan_slice() to make it easy to
thin samples from each parameter in a stanfit object.
}
\examples{
rats <- rats_example(nCores = 1)

rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_thin_n(30)
  
rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_thin_frac(0.5)

rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_thin_n(30,inc_warmup = FALSE)

}
\seealso{
Other filtering: \code{\link{stan_filter}},
  \code{\link{stan_slice}}
}
\concept{filtering}
