% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats_example.R
\name{rats_example}
\alias{rats_example}
\title{Rats model}
\source{
\href{https://www.mrc-bsu.cam.ac.uk/wp-content/uploads/WinBUGS_Vol1.pdf}{Data Description},
\href{https://github.com/stan-dev/example-models/tree/master/bugs_examples/vol1/rats}{Data and Script}
}
\usage{
rats_example(seed = 123, nCores = pmin(parallel::detectCores(), 4))
}
\arguments{
\item{seed}{numeric, seed for \link[rstan:stan]{rstan}, Default: 123}

\item{nCores}{numeric, Maximum number of cores to use, Default: pmin(parallel::detectCores(),4)}
}
\value{
stanfit object
}
\description{
Run Rats data model
}
\examples{
\donttest{
  require(shredder)
  rats <- rats_example(nCores = 1)
  rats
}
}
\seealso{
\link[rstan:stan]{rstan}
\link[parallel:detectCores]{detectCores}

Other examples: 
\code{\link{rat_data}}
}
\concept{examples}
