% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_filter.R
\name{stan_filter}
\alias{stan_filter}
\title{Return post-warmup samples with matching conditions}
\usage{
stan_filter(object, ..., chain = 1)
}
\arguments{
\item{object}{stanfit object}

\item{\dots}{Logical predicates defined in terms of the parameters in object}

\item{chain}{numeric, chain to apply filter predicated on, Default: 1}
}
\value{
stanfit object
}
\description{
Use stan_filter() to choose indicies of samples across parameters
where conditions are true.
}
\examples{
rats <- rats_example()

rats\%>\%
  stan_select(mu_alpha,mu_beta)
  
rats\%>\%
  stan_select(mu_alpha,mu_beta)\%>\%
  stan_filter(mu_beta < 6)
  
}
\seealso{
Other filtering: \code{\link{stan_sample_frac}},
  \code{\link{stan_slice}}
}
\concept{filtering}
