% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_filter.R
\name{stan_filter}
\alias{stan_filter}
\title{Return post-warmup samples with matching conditions}
\usage{
stan_filter(object, ..., permuted = TRUE)
}
\arguments{
\item{object}{stanfit object}

\item{\dots}{Logical predicates defined in terms of the parameters in object}

\item{permuted}{A logical scalar indicating whether the draws after the warmup
period in each chain should be permuted and merged, Default: TRUE}
}
\value{
stanfit object
}
\description{
Use stan_filter() to choose indicies of samples across parameters
where conditions are true.
}
\details{
\itemize{
\item If no elements are returned for any chain then \code{NULL} is returned with a warning.
\item If there is a chain that results in no samples then the chain is dropped with a warning.
\item If permuted is FALSE then uneven chains may be returned depending on the result of the filter.
\item To comply with \link[rstan:extract]{extract} with \code{permuted=TRUE} chains in which chaines are
assumed to be of equal size.
\itemize{
\item Chain size returned is the length of the shortest filtered chain.
}
}
}
\examples{
rats <- rats_example(nCores = 1)

rats\%>\%
  stan_select(mu_alpha,mu_beta)
  
rats\%>\%
  stan_select(mu_alpha,mu_beta)\%>\%
  stan_filter(mu_beta < 6)
  
rats\%>\%
  stan_select(mu_alpha,mu_beta)\%>\%
  stan_filter(mu_beta < 6, permuted = FALSE)
  
}
\seealso{
Other filtering: \code{\link{stan_sample_n}},
  \code{\link{stan_slice}}
}
\concept{filtering}
