% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partials.R
\name{stan_contains}
\alias{stan_contains}
\alias{stan_starts_with}
\alias{stan_ends_with}
\title{Partial match functions for stanfit objects}
\usage{
stan_contains(match, ignore.case = TRUE, pars = peek_pars())

stan_starts_with(match, ignore.case = TRUE, pars = peek_pars())

stan_ends_with(match, ignore.case = TRUE, pars = peek_pars())
}
\arguments{
\item{match}{character, pattern to search for}

\item{ignore.case}{logical to ignore the case, Default: TRUE}

\item{pars}{list of pars to search in, Default: peek_pars()}
}
\value{
character
}
\description{
Partial match functions used in stan_select
}
\seealso{
Other subsetting: \code{\link{stan_select}}
}
\concept{subsetting}
