% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_sample.R
\name{stan_sample_n}
\alias{stan_sample_n}
\alias{stan_sample_frac}
\title{Sample n post-warumps samples from a stanfit object}
\usage{
stan_sample_n(object, size, weight = NULL, inc_warmup = TRUE)

stan_sample_frac(object, size, weight = NULL, inc_warmup = TRUE)
}
\arguments{
\item{object}{stanfit object}

\item{size}{numeric, for \link[=stan_sample_n]{shredder} size of sample, for
\link[=stan_sample_frac]{shredder} fraction of samples to sample.}

\item{weight, }{a vector of probability weights for obtaining the elements of
the vector being sampled.}

\item{inc_warmup}{logical, include warmup in output, Default: TRUE}
}
\value{
stanfit
}
\description{
This is a wrapper around sample() to make it easy to select
random samples from each parameter in a stanfit object.
}
\examples{
rats <- rats_example(nCores = 1)

rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_sample_n(30)
  
rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_sample_frac(0.5)

rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_sample_n(30,inc_warmup = FALSE)

}
\seealso{
Other filtering: \code{\link{stan_filter}},
  \code{\link{stan_slice}}
}
\concept{filtering}
