% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_select.R
\name{stan_select}
\alias{stan_select}
\title{Select parameters by name}
\usage{
stan_select(object, ...)
}
\arguments{
\item{object}{a stanfit object}

\item{...}{One or more unquoted expressions separated by commas}
}
\value{
stanfit object
}
\description{
Choose variables from a stanfit object
}
\examples{
set.seed(123)
rats <- rats_example()

rats\%>\%stan_select(mu_beta)

rats\%>\%stan_select(mu_beta,beta)

rats\%>\%stan_select(!!!rlang::syms(c('mu_beta','mu_alpha')))

}
\seealso{
\link[rlang:quotation]{quotation}, \link[rlang:quo_label]{quo_label}, \link[purrr:map]{map}
}
