% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_select.R
\name{stan_select}
\alias{stan_select}
\title{Select parameters by name}
\usage{
stan_select(object, ...)
}
\arguments{
\item{object}{a stanfit object}

\item{...}{One or more unquoted expressions separated by commas}
}
\value{
stanfit object
}
\description{
Choose variables from a stanfit object
}
\examples{
rats <- rats_example(nCores = 1)

rats\%>\%stan_select(mu_beta)

rats\%>\%stan_select(mu_beta,beta)

rats\%>\%stan_select(!!!rlang::syms(c('mu_beta','mu_alpha')))

}
\seealso{
\link[rlang:nse-defuse]{quotation}, \link[rlang:quo_label]{quo_label}, \link[purrr:map]{map}

Other subsetting: 
\code{\link{stan_contains}()}
}
\concept{subsetting}
