% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_slice.R
\name{stan_slice}
\alias{stan_slice}
\title{Choose post-warumps samples by position}
\usage{
stan_slice(object, ..., inc_warmup = TRUE)
}
\arguments{
\item{object}{stanfit object}

\item{...}{Integer samples values}

\item{inc_warmup}{logical, include warmup in output, Default: TRUE}
}
\value{
stanfit object
}
\description{
Choose post-warumps samples by their ordinal
position from each parameter in a stanfit object
}
\examples{
\donttest{
rats <- rats_example(nCores = 1)

rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_slice(1:30)

rats\%>\%
  stan_select(mu_alpha)\%>\%
  stan_slice(1:30,inc_warmup = FALSE)
}
}
\seealso{
Other filtering: 
\code{\link{stan_filter}()},
\code{\link{stan_thin_n}()}
}
\concept{filtering}
