% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_split.R
\name{stan_split}
\alias{stan_split}
\title{split stanfit object into smaller pieces}
\usage{
stan_split(object, ncut = 10, inc_warmup = TRUE)
}
\arguments{
\item{object}{stanfit}

\item{ncut}{numeric, number of smaller objects to create, Default: 10}

\item{inc_warmup}{logical, include warmup in the smaller objects, Default: TRUE}
}
\value{
list of stanfits objects
}
\description{
Split stanfit post-warmup samples into smaller objects to
facilitate parallel post processing.
}
\examples{
rats <- rats_example(nCores = 1)

rats\%>\%
  stan_split(ncut = 4)
  
rats\%>\%
  stan_split(ncut = 4,inc_warmup = FALSE)

}
\seealso{
\link[shredder:stan_slice]{stan_slice}

Other utility: \code{\link{stan_names}}
}
\concept{utility}
